requirements.txt
anthropic>=0.84.0
mcp>=1.26.0
python-dotenv>=0.21.0
pandas>=2.2.2
pyyaml>=6.0.1
fastembed>=0.4.0
asyncpg>=0.29.0
